double instead of unsigned

+++ test.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-    vector<vector<unsigned> > vec;
+    vector<vector<double> > vec;
     image_read("usFlag.png", 425, 276, vec);
     image_write("image.png", vec);
     return 0;
